refactor: app wave-15 Slice B widget splits and 300-line ratchets
Split near-cap game/catalog widgets and lower game-widgets plus new
app/lib/widgets catalog gates to 300 so AC3/AC4 hold without UI drift.
