+++ SPEC/program/repo-lint.md
@@ -18,7 +18,7 @@
 
 **Implementation status (GitHub #1912, dart file size):** `repo.dart_file_non_comment_line_size` enforces a universal **1000** non-comment-line cap for scanned hand-written Dart; it skips only **generated whole paths** (suffixes `.g.dart`, `.freezed.dart`, `.mocks.dart`, `.gen.dart`, and generated Flutter l10n path prefixes under `app/lib/l10n/gen/`). Per-rule contract: `SPEC/program/dart-file-non-comment-line-size.md`.
 
-**Implementation status (GitHub #1912, game widgets file size):** `repo.game_widgets_file_size` enforces a universal **400** physical-line cap for every `.dart` file under `app/lib/features/game/widgets/**` with no keyed waivers (wave-14 #4305 Slice C, down from 700). Per-rule contract: `SPEC/program/game-widgets-file-size.md`.
+**Implementation status (GitHub #1912 / #4352, game widgets file size):** `repo.game_widgets_file_size` enforces a universal **300** physical-line cap for every `.dart` file under `app/lib/features/game/widgets/**` with no keyed waivers (wave-15 #4352 Slice B, down from wave-14 **400**). Per-rule contract: `SPEC/program/game-widgets-file-size.md`. Sibling `repo.app_catalog_widgets_file_size` enforces the same **300** cap under `app/lib/widgets/**` (Refs #4352 AC4; `SPEC/program/app-catalog-widgets-file-size.md`).
 
 **Implementation status (GitHub #1912, tech id literals):** `repo.tech_id_constants` treats the string values of every `const String kTechId*` in `packages/colonizethis_data/lib/src/tech_ids.dart` as the canonical tech-id set (replacing the prior `tech_catalog.dart` map-key scrape). `tech_catalog.dart` uses those same constants for map keys and `TechDefinition` ids so the catalog stays aligned with the checker.
 
@@ -63,7 +63,8 @@
 | `tool/check_workspace_outdated_latest_direct.dart` | Enforce that audited direct dependency rows are at `Latest` whenever `resolvable == latest` (repo root + every workspace member using `sdk: flutter`); rule `repo.workspace_outdated_latest_direct` |
 | `tool/check_no_flame_in_widgets.dart` | Disallow direct `package:flame/*` **import** and **export** lines (single- or double-quoted URI) under `app/lib/widgets/**`; rule `repo.no_flame_in_widgets` |
 | `tool/check_no_screen_in_game_widgets.dart` | Disallow `*_screen.dart` files under `app/lib/features/game/widgets/**`; rule `repo.no_screen_in_game_widgets` |
-| `tool/check_game_widgets_file_size.dart` | Enforce `app/lib/features/game/widgets/**` Dart files at **400 physical lines or fewer** — see `SPEC/program/game-widgets-file-size.md`; rule `repo.game_widgets_file_size` (Refs #4305 Slice C) |
+| `tool/check_game_widgets_file_size.dart` | Enforce `app/lib/features/game/widgets/**` Dart files at **300 physical lines or fewer** — see `SPEC/program/game-widgets-file-size.md`; rule `repo.game_widgets_file_size` (Refs #4352 Slice B) |
+| `tool/check_app_catalog_widgets_file_size.dart` | Enforce `app/lib/widgets/**` hand-written Dart files at **300 physical lines or fewer** — see `SPEC/program/app-catalog-widgets-file-size.md`; rule `repo.app_catalog_widgets_file_size` (Refs #4352 Slice B / AC4) |
 | `tool/check_logic_test_file_size.dart` | Enforce `packages/colonizethis_logic/test/**` Dart files at **400 physical lines or fewer** (physical line count); rule `repo.logic_test_file_size` runs **full-tree** in `ct_repo_lint` and narrows to changed files when an incremental baseline is supplied (`CT_REPO_LINT_BASE_SHA` in CI). Originally PR-incremental only against the #2216 debt; transitioned to full-tree enforcement once the residual `test/**` violations were cleared (GitHub #2288). The dedicated `quality.yml` step for changed logic tests remains as a fast early signal on PRs. |
 | `tool/check_logic_test_tree_loc.dart` | Enforce total physical LOC under `packages/colonizethis_logic/test/**` at **≤5,800** (post-#4090 orphan purge ratchet); rule `repo.logic_test_tree_loc`. |
 | `tool/check_logic_test_orphan_basenames.dart` | Forbid reintroduction of the thirteen purged domain collision basenames under `packages/colonizethis_logic/test/**` (Refs #4090 AC3/AC5); rule `repo.logic_test_orphan_basenames`. |
