+++ tool/check_game_widgets_file_size.dart
@@ -1,15 +1,15 @@
 // Physical line limit for game feature widgets (`repo.game_widgets_file_size`).
-// SPEC: SPEC/program/game-widgets-file-size.md
+// SPEC: SPEC/program/game-widgets-file-size.md (wave-15 #4352 Slice B).
 import 'dart:convert';
 import 'dart:io';
 
 import 'package:path/path.dart' as p;
 
-/// Ratchet ceiling for game feature widgets (wave-14 #4305 Slice C).
-const int gameWidgetsFileSizeCeiling = 400;
+/// Ratchet ceiling for game feature widgets (wave-15 #4352 Slice B).
+const int gameWidgetsFileSizeCeiling = 300;
 
 /// PR-blocking structural check: files under
-/// `app/lib/features/game/widgets/**` must stay at or below 400 physical lines.
+/// `app/lib/features/game/widgets/**` must stay at or below 300 physical lines.
 int runCheckGameWidgetsFileSize(
   String repoRoot, {
   void Function(String line)? info,
@@ -48,12 +48,16 @@ int runCheckGameWidgetsFileSize(
   }
 
   if (violations.isEmpty) {
-    logI('check_game_widgets_file_size: no violations found.');
+    logI(
+      'check_game_widgets_file_size: no violations found '
+      '(ceiling $gameWidgetsFileSizeCeiling; Refs #4352).',
+    );
     return 0;
   }
 
   logE(
-    'check_game_widgets_file_size: found ${violations.length} violation(s) under app/lib/features/game/widgets:',
+    'check_game_widgets_file_size: found ${violations.length} violation(s) '
+    'under app/lib/features/game/widgets:',
   );
   for (final violation in violations) {
     logE(' - $violation');
